requirements.txt
streamlit==1.10.0
Pillow==9.0.1
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
